% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_startup_data.R
\name{deleteStartupData}
\alias{deleteStartupData}
\title{Delete data for an interval after each power-up event}
\usage{
deleteStartupData(
  x,
  gapTime = 60,
  deleteTime = 20,
  stream = "payload1",
  debug = getOption("gliderDebug", default = 0)
)
}
\arguments{
\item{x}{an [oceglider-class] object, e.g. as read by
`read.read.glider.seaexplorer.raw()` or similar
functions.}

\item{gapTime}{numeric value, in seconds, used to infer power-on
events.  There is always one event at the start of the time series. In
addition, any time differences that exceed `startupCriterion` seconds
are considered to be power-off events, so the first time after such
an interval is taken to be a power-on event.}

\item{deleteTime}{numeric value indicating the time interval,
in seconds, of data that will be removed.}

\item{stream}{character value naming the stream to use, if the object
is multi-streamed, meaning that the `data` slot contains multiple data frames,
one for the glider (named `glider`) another for the first payload (named
`"payload"`), and possibly other payloads. The value of `stream` is ignored
unless the object's `metadata` slot contains an item named `dataAreStreamed`
and that item is TRUE.}

\item{debug}{an integer, less than or equal to zero for no message, and
positive to print information about the processing.}
}
\description{
This trims the spurious data that instruments can produce
for a short time interval after each power-on events.
}
\author{
Dan Kelley
}
