% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\name{download.glider}
\alias{download.glider}
\title{Download and Cache a Glider File}
\usage{
download.glider(url, pattern, destdir = ".", debug = 0)
}
\arguments{
\item{url}{Character value providing the web location of the
server.}

\item{pattern}{Character value indicating a pattern for
the name(s) of the desired file(s).}

\item{destdir}{Character value indicating the directory in which
to save the downloaded data.}

\item{debug}{Integer indicating the debugging level; 0 for quiet
action and higher values for more indications of the processing
steps.}
}
\value{
Either a vector of character values for the names of the
desired files (whether they were downloaded or already present
locally), or \code{NULL} if the server had no files
matching the indicated pattern.
}
\description{
This function visits a server that holds glider data,
looking for one or more files matching a given pattern. If such
files are found on the server, then those files that do
not already exist locally are downloaded
using \code{\link[utils]{download.file}}, which produces
messages to indicate the course of its progress. If no
files on the server match the pattern, then the return value
is \code{NULL}; otherwise, it is a vector of character values
giving the names of the files that have been downloaded.
}
\examples{
# Download and read yo 200 of glider SEA024 on mission 32.
\dontrun{
url <- "ftp://ftp.dfo-mpo.gc.ca/glider/realData/SEA024/M32"
yofiles <- download.glider(url, "\\\\.200\\\\.gz$", destdir="~/data/glider")
if (2 == length(yofiles)) {
    yo <- read.glider.seaexplorer(yofiles)
}
summary(yo)
}

}
\seealso{
Other functions to download data: \code{\link{download.glider.seaexplorer}},
  \code{\link{download.glider.slocum}}
}
\author{
Dan Kelley
}
\concept{functions to download data}
