% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaexplorer.R
\name{download.glider.seaexplorer}
\alias{download.glider.seaexplorer}
\title{Download and Cache a Seaexplorer Glider File [deprecated]}
\usage{
download.glider.seaexplorer(url = "ftp://ftp.dfo-mpo.gc.ca/glider",
  stream = "realData", glider = "SEA024", mission = "M32",
  type = "pld1", yo = "?", debug = 0)
}
\arguments{
\item{url}{Character value indicating the base URL for the
server. If this is not given as in the default, it seems
fairly likely that this function will fail.
If \code{url} is \code{"?"}, the default is printed.}

\item{stream}{Character value indicating the data stream. This
should probably be either \code{"realData"} or \code{"delayedData"}.
If \code{stream} is \code{"?"}, the data repository will
be examined, and a message will be printed about possible values.}

\item{glider}{Name Character value indicating the name of the
glider. This is probably in a form like \code{"SEA021"}.
If \code{glider} is \code{"?"}, the data repository will
be examined, and a message will be printed about possible values.}

\item{mission}{Character value indicating the name of the
mission. This is probably in a form like \code{"M32"}.
If \code{missionName} is \code{"?"}, the data repository will
be examined, and a message will be printed about possible values.}

\item{type}{Character value, either \code{"pld1"} or \code{"gli"}.}

\item{yo}{Numerical value indicating the yo number,
or the character value \code{"?"}.  In the second case,
the data repository will be examined, and the return
value will be a vector of \code{yo} numbers that can
be retrieved by the server.}

\item{debug}{Integer indicating the debugging level; 0 for quiet
action and higher values for more indications of the processing
steps.}
}
\value{
Either a character vector of file names or (if \code{yo=="?"})
a numerical vector of possible \code{yo} values for the indicated
server.
}
\description{
This function assumes some knowledge of the data being sought,
but if some parameters are set to \code{"?"}, the server
will be interrogated to try to determine possible choices.
}
\examples{
\dontrun{
# Download and read a file (default server, mission, etc)
yo <- 200
url <- "ftp://SERVER/PATH"
files <- download.glider(url, paste("\\\\.", yo, "\\\\.gz$", sep=""), debug=1)
yo2 <- read.glider.seaexplorer.realtime(files)
# Download (or use cache for) a set files
download.glider.seaexplorer(yo=download.glider.seaexplorer(yo="?"))
}

}
\seealso{
Other functions for seaexplorer gliders: \code{\link{read.glider.seaexplorer.delayed}},
  \code{\link{read.glider.seaexplorer.realtime}}

Other functions to download data: \code{\link{download.glider.slocum}},
  \code{\link{download.glider}}
}
\author{
Dan Kelley
}
\concept{functions for seaexplorer gliders}
\concept{functions to download data}
