% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slocum.R
\name{download.glider.slocum}
\alias{download.glider.slocum}
\title{Download and Cache a Slocum Glider File}
\usage{
download.glider.slocum(mission = "m80", year = 2017, month = 12,
  day = 16, item = "view_sci_water.csv",
  server = "http://gliders.oceantrack.org/data/slocum", destdir = ".",
  destfile, force = FALSE, dryrun = FALSE,
  debug = getOption("gliderDebug", 0))
}
\arguments{
\item{mission}{Character value indicating the mission name.}

\item{year}{Numeric value indicating the year of the mission start.}

\item{month}{Numeric value indicating the month of the mission start.}

\item{day}{Numeric value indicating the day of the mission start.}

\item{item}{Character value indicating the type of data sought.}

\item{server}{String specifying the online server.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/glider}
to hold glider files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{gliderDebug}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{glider} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
A character value indicating the filename of the result; if
there is a problem, the result will be the empty string.
}
\description{
If the file is already present in \code{destdir}, then it is not
downloaded again. The default \code{destdir} is the present directory,
but it probably makes more sense to use something like \code{"~/data/glider"}
to make it easy for scripts in other directories to use the cached data.
The file is downloaded with \code{\link{download.file}}.
}
\examples{
\dontrun{
## The download takes several seconds.
library(oceanglider)
gfile <- download.glider.slocum(destddir="~/data/glider")
}
}
\seealso{
Other functions for slocum gliders: \code{\link{read.glider.slocum}}

Other functions to download data: \code{\link{download.glider.seaexplorer}},
  \code{\link{download.glider}}
}
\concept{functions for slocum gliders}
\concept{functions to download data}
