% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{gliderDebug}
\alias{gliderDebug}
\title{Print a debugging message}
\usage{
gliderDebug(debug = 0, ..., unindent = 0)
}
\arguments{
\item{debug}{an integer, less than or equal to zero for no message, and
greater than zero for increasing levels of debugging.  Values greater than 4
are treated like 4.}

\item{\dots}{items to be supplied to \code{\link[=cat]{cat()}}, which does the
printing.  Almost always, this should include a trailing newline.}

\item{unindent}{Number of levels to unindent, e.g. for start and end lines
from a called function.}
}
\description{
Many glider functions decrease the \code{debug} level by 1 when they call other
functions, so the effect is a nesting, with more space for deeper function
level.
}
\author{
Dan Kelley
}
