% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\name{gliderTrim}
\alias{gliderTrim}
\title{Trim a glider Object}
\usage{
gliderTrim(x, method)
}
\arguments{
\item{x}{A \code{glider} object, i.e. one inheriting from
\code{\link{glider-class}}.}

\item{method}{An expression indicating how to subset \code{x}. See
\dQuote{Details}.}
}
\value{
A \code{\link{glider-class}} object that
has been trimmed to contain only the data specified by
\code{subset}.
}
\description{
Return a trimmed version of a glider object.
}
\details{
At the moment, this only works for SeaExplorer data (i.e. cases in which
\code{x[["type"]]=="seaexplorer"}).

The permitted values of \code{method} are:
\itemize{

\item \code{"ascending"}, which retains only \code{glider}
data entries for which the \code{NavState} equals 117, and
only \code{payload} data entries for which
\code{NAV_RESOURCE} is 117.

\item \code{"descending"}, which retains only \code{glider}
data entries for which the \code{NavState} equals 100, and
only \code{payload} data entries for which
\code{NAV_RESOURCE} is 100.

}
}
\section{Caution}{

This function may be subsumed into \code{\link{subset,glider-method}}, because it
does similar things, and users are more likely to guess the name of the latter.
}

\examples{
files <- system.file("extdata/seaexplorer/realtime",
                     c("sea024.32.gli.sub.200.gz",
                       "sea024.32.pld1.sub.200.gz"), package="oceanglider")
d <- read.glider.seaexplorer.realtime(files)
summary(gliderTrim(d, "ascending"))
summary(gliderTrim(d, "descending"))

}
\author{
Dan Kelley
}
