% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\name{handleFlagsInternalOceanglider}
\alias{handleFlagsInternalOceanglider}
\title{Low-level function to handle flags (temporary code)}
\usage{
handleFlagsInternalOceanglider(object, flags, actions, debug)
}
\arguments{
\item{object}{An \code{oceanglider} object, i.e. an object inheriting
from \code{\link{oceanglider-class}}.}

\item{flags}{A \code{\link{list}} that associates integer values}

\item{actions}{A character vector, which is lengthened to match
the length of \code{flags}. The most common value is \code{"NA"},
which means to set flaggd values to the missing-value cod, \code{NA}.}

\item{debug}{An integer specifying the debugging level, with value
\code{0} meaning to act silently, and higher values meaning to print
some debugginf information.}
}
\description{
**Important note.** This function is provided only because the CRAN version
of the oce package does not export \code{handleFlagsInternal}. Once a
the CRAN oce is updated, this function will be removed.
}
\author{
Dan Kelley
}
