% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\docType{methods}
\name{initialize,glider-method}
\alias{initialize,glider-method}
\title{Class for Glider Objects}
\usage{
\S4method{initialize}{glider}(.Object, filename)
}
\arguments{
\item{.Object}{Object.}

\item{filename}{Name of file.}
}
\description{
Class for Glider Objects
}
\author{
Dan Kelley
}
