% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaexplorer.R
\name{navStateCodes}
\alias{navStateCodes}
\title{Possible navState values of a glider object}
\usage{
navStateCodes(g)
}
\arguments{
\item{g}{Either a character string or glider object. If it is a string,
then it is the type of glider, which in the present version of the
function must be \code{"seaexplorer"}. If it is
a glider object, then the value of \code{navStateCodes} in the \code{metadata}
slot of that object is returned, if that exists, or else the \code{type}
item in the \code{metadata} slot is used to determine the type, as
in the case with \code{g} being a character string.}
}
\value{
A list of integers defining the navigation state, each
given a brief name as indicated in the \dQuote{Details} section.
}
\description{
This function provides names for the numerical \code{navState} codes
used by various gliders, with the aim of making it easier to
write self-explantory code (see \dQuote{Examples}).
}
\details{
The numerical values for \code{seaexplorer} are as follows.
\tabular{lll}{
\strong{Name} \tab \strong{Value} \tab \strong{Description}\cr
\code{not_navigating}       \tab 105 \tab glider is being set up\cr
\code{surfacing}            \tab 115 \tab nearing the surface\cr
\code{at_surface}           \tab 116 \tab at the surface, acquiring GPS and transmitting data\cr
\code{inflecting_downwards} \tab 110 \tab ballast being adjusted to cause descent\cr
\code{descending}           \tab 100 \tab ballast causing descent\cr
\code{inflecting_upwards}   \tab 118 \tab ballast being adjusted to cause ascent\cr
\code{ascending}            \tab 117 \tab ballast causing ascent\cr
}
Note that the downward portions of profiles are roughly signalled by several
\code{inflecting_downwards} codes followed by \code{descending}
codes, while the upward portions have \code{inflecting_upwards} codes

The numerical values for type \code{slocum} are as follows. (These
are(defined as \code{m_depth_state} in the \code{slocum} documentation;
see pages 1-24 of reference 1.)
\tabular{lll}{
\strong{Name}   \tab \strong{Value} \tab \strong{Description}\cr
\code{ignore}   \tab        99 \tab              - \cr
\code{hover}    \tab         3 \tab              - \cr
\code{climbing} \tab         2 \tab              - \cr
\code{diving}   \tab         1 \tab              - \cr
\code{surface}  \tab         0 \tab              - \cr
\code{none}     \tab        -1 \tab              - \cr
}
}
\examples{
# Use codes to identify upcasts, at least roughly (note the stray points)
directory <- system.file("extdata/seaexplorer/raw", package="oceGlider")
g <- read.glider.seaexplorer.delayed(directory)
ns <- navStateCodes(g)
plot(g, which="p")
ga <- subset(g, navState == ns$ascending)
points(ga[["time"]], ga[["pressure"]], col=3, pch=20)
giu <- subset(g, navState == ns$inflecting_upwards)
points(giu[["time"]], giu[["pressure"]], col=2, pch=20)
mtext(" red=inflecting_upwards; green=ascending", side=3, line=-1, adj=0)

}
\references{
1.Teledyne Webb Research. \emph{Slocum G2 Glider Operators Manual}, January 2012.
\url{https://gliderfs2.coas.oregonstate.edu/gliderweb/docs/slocum_manuals/Slocum_G2_Glider_Operators_Manual.pdf}.
}
\author{
Dan Kelley
}
