% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\docType{package}
\name{oceanglider-class}
\alias{oceanglider-class}
\alias{oceanglider-class-package}
\title{oceanglider: A Package for Processing Ocean Glider Data}
\description{
This package was written with two particular dataset types
in mind, from SeaExplorer and Slocum devices. There is a good
chance that the functions provided here (a) will fail on
other types and (b) function names and arguments will change
as more datasets are examined by the author.
}
\details{
The downloading functions require a base URL, and the default argument is to
use values stored as options.  For the author, this is done by putting the following
in the top-level \code{.Rprofile} file:

\preformatted{
options(slocumURL="")
options(seaexplorerURL="")
}
}
