% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/netcdf.R
\docType{package}
\name{oceglider}
\alias{oceglider}
\title{oceglider: A Package for Processing Ocean Glider Data}
\description{
This package was written with two particular dataset types
in mind, from SeaExplorer and Slocum devices. There is a good
chance that the functions provided here (a) will fail on
other types and (b) function names and arguments will change
as more datasets are examined by the author.

This package was written with two particular dataset types
in mind, from SeaExplorer and Slocum devices. There is a good
chance that the functions provided here (a) will fail on
other types and (b) function names and arguments will change
as more datasets are examined by the author.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dankelley.github.io/oceglider/}
  \item Report bugs at \url{https://github.com/dankelley/oceglider/issues}
}


Useful links:
\itemize{
  \item \url{https://dankelley.github.io/oceglider/}
  \item Report bugs at \url{https://github.com/dankelley/oceglider/issues}
}

}
\author{
\strong{Maintainer}: Dan Kelley \email{dan.kelley@dal.ca} (\href{https://orcid.org/0000-0001-7808-5911}{ORCID})

Authors:
\itemize{
  \item Clark Richards \email{clark.richards@gmail.com} (\href{https://orcid.org/0000-0002-7833-206X}{ORCID})
  \item Chantelle Layton \email{layton.chantelle@gmail.com} (\href{https://orcid.org/0000-0002-3199-5763}{ORCID})
}

}
\keyword{internal}
