% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\docType{methods}
\name{plot,glider-method}
\alias{plot,glider-method}
\title{Plot a glider Object}
\usage{
\S4method{plot}{glider}(x, which, ...)
}
\arguments{
\item{x}{A \code{glider} object, i.e. one inheriting from \code{\link{glider-class}}.}

\item{which}{Integer or character value indicating which type of plot to produce
(see \dQuote{Details}).}

\item{...}{Ignored in the present version.}
}
\description{
The type of plot depends on the value of the \code{which} argument:
\itemize{
\item \code{which=1}: plot pressure versus time
\item \code{which=2}: plot temperature versus time
\item \code{which=3}: plot salinity versus time
}
}
