% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioos.R
\name{read.glider.netcdf.ioos}
\alias{read.glider.netcdf.ioos}
\title{Read a glider file in netcdf IOOS format}
\usage{
read.glider.netcdf.ioos(file, debug)
}
\arguments{
\item{file}{Name of a netcdf file.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. If this is not provided, then
the value of \code{\link{getOption}("gliderDebug",0)} is used.
The printing is done by a call to \code{\link{gliderDebug}}.
Setting \code{debug=0} turns off this form of debugging, while
higher values may yield more information, depending on the
function. If one \code{glider} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
A glider object, i.e. one inheriting from \linkS4class{glider}.
(This class inherits from \link[oce:oce-class]{oce::oce} in the
\CRANpkg{oce} package.)
}
\description{
\strong{This is a provisional function, written to handle provisional
netcdf files created from the python \code{pyglider} processing chain
as well as files downloaded from the Glider IOOS ERDDAP site.}
}
\details{
The data are copied directly from the file, except that \code{time}
is converted from an integer to a POSIX time. Variable names containing
underscores are renamed as e.g. \code{profile_direction}
to \code{profileDirection}, although the \code{\link{[[,glider-method}}
mechanism works with either name, e.g. if \code{g} is a glider object, then
\code{g[["profileDirection"]]} and
\code{g[["profile_direction"]]} give the same result.
}
\seealso{
Other functions to read glider data: 
\code{\link{read.glider.netcdf}()},
\code{\link{read.glider.seaexplorer.delayed}()},
\code{\link{read.glider.seaexplorer.realtime}()},
\code{\link{read.glider.slocum}()}
}
\author{
Chantelle Layton and Dan Kelley
}
\concept{functions to read glider data}
