% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glider.R
\name{read.glider}
\alias{read.glider}
\title{Read a Glider file}
\usage{
read.glider(name)
}
\arguments{
\item{name}{Character value giving the name of the file}
}
\value{
An oce object holding the data.
}
\description{
Read a Glider file
}
\examples{
\dontrun{
g <- read.glider("~/data/glider/m80_2017-12-16_view_sci_water.csv")
}
}
\author{
Dan Kelley
}
