% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaexplorer_raw.R
\name{read.glider.seaexplorer.raw}
\alias{read.glider.seaexplorer.raw}
\title{Read Raw SeaExplorer Data}
\usage{
read.glider.seaexplorer.raw(
  directory,
  pattern = "pld1.raw",
  yo,
  level = 1,
  interpolateToCTD = TRUE,
  rename = TRUE,
  progressBar = interactive(),
  missingValue = 9999,
  debug = getOption("gliderDebug", default = 0)
)
}
\arguments{
\item{directory}{The directory in which the delayed-mode SeaExplorer files are located.}

\item{pattern}{a character value used to find files to read. The
default, \code{"pld1"}, will match files with that string in the name. The
best choices for this depends on the setup of files. The search
for files is done with \code{list.files(directory, pattern)}, and so
it can be a good idea to try that first, to learn how to specify
the desired files.  Also, try calling with \code{debug=1} to get some
indication of the files that get read or with \code{debug=2} to get
a full listing.}

\item{yo}{A numeric value (or vector) specifying the yo numbers to read. If
this is not provided, \code{\link[=read.glider.seaexplorer.raw]{read.glider.seaexplorer.raw()}} will read all yo
numbers for which files are present in \code{dir}.}

\item{level}{A numeric value specifying the processing level, 0 or
\enumerate{
\item See Details.
}}

\item{interpolateToCTD}{A logical indicating whether all sensors should be
interpolated to the CTD times to obtain a common time base, or whether all
sensors should simply be interpolated for all time stamps (which was the
default behaviour before 2019-12-08)}

\item{rename}{an indication of how (or if) to rename variables. This is
needed for most practical work in the oce package, which expects
standardized names, such as \code{"temperature"}, as opposed to the names stored
in glider files. There are three choices for \code{rename}. (a) It can be
logical, with TRUE (the default) meaning to use names as defined in
\code{system.file("extdata/dictionaries/seaexplorerDict.csv",package="oceglider")}
or FALSE, meaning not to rename variables. (b) It can be the name of a CSV
file that is in the same format as the file above-named file.  (c)
It can be a two-column data frame in which column 1 holds
variable names in the glider file and column 2 holds the corresponding
names to be used in the return value.}

\item{progressBar}{a logical value that controls whether to indicate the
progress made in reading and interpreting the data.  This can be useful,
since the work can be slow. The default is to show progress in
interactive sessions, but not in scripts.}

\item{missingValue}{numeric value that indicates bad data. Any data items
equaling this value are converted to NA. The default is 9999. To avoid
changing values to NA, call the function with \code{missingValue=NULL}.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. If this is not provided, then
the value of \code{\link{getOption}("gliderDebug",0)} is used.
The printing is done by a call to \code{\link{gliderDebug}}.
Setting \code{debug=0} turns off this form of debugging, while
higher values may yield more information, depending on the
function. If one \code{glider} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\description{
Reads raw CSV files produced by a SeaExplorer glider. This function can
output either "Level 0" or "Level 1" type data. Level 0 is simply the raw
data as written in the CSV files with very processing done except
renaming variables to the oce convention (e.g. \code{"GPCTD_CONDUCTIVITY"}
is renamed as \code{"conductivity"}), converting timestamp strings
into times, and converting longitude and latitude into decimal
degrees.
}
\details{
Level 1 processing performs a number of steps to give an
"analysis ready" dataset, including

\enumerate{

\item If the dataset lacks a dead-reckoning column, surface longitude and
latitude readings are interpolated with respect to time, to provide a crude
estimate subsurface positions.

\item Interpolation, depending on the value of \code{interpolateToCTD}. If
\code{interpolateToCTD} is \code{TRUE}, then any "extra" sensors are interpolated
to the times for which there is CTD data. Otherwise, NAs for \emph{all}
the sensors are interpolated to a common time, corresponding to the
raw time stamps output from the various sensors. A caution -- this
will produce an apparent "upsampling" of each sensor, so that the
apparent sample rate is higher. For example, if a Wetlabs FLBBCD
sensor sampled, but there is no corresponding GP-CTD sample from
the same time, the CTD parameters will be interpolated from the
ones before and after. This has the disadvantage of interpolating
values that were not measured, but has the advantage of assigning
pressures to values measured by sensors not integrated into the CTD
(e.g. Wetlabs FLBBCD, Rinko O2). Following the interpolation, any
rows with duplicated times are removed.

\item calculation of Practical salinity from conductivity, temperature
and pressure using \code{\link[oce:swSCTp]{oce::swSCTp()}}.

}

In any case, a flag scheme is set up according to the IOOS classification system (see
Table 2 of reference 1), as follows.

\tabular{llll}{
\strong{Name}   \tab \strong{Value} \tab \strong{IOOS Name}          \tab \strong{Description}\cr
\code{pass}          \tab 1              \tab Pass                        \tab Data has passed quality control (QC) tests\cr
\code{not_evaluated} \tab 2              \tab Not Evaluated               \tab Data has not been QC tested\cr
\code{suspect}       \tab 3              \tab Suspect or of High Interest \tab Data is considered to be of suspect or high interest\cr
\code{fail}          \tab 4              \tab Fail                        \tab Data is considered to have failed on one or more QC tests\cr
\code{missing}       \tab 9              \tab Missing Data                \tab Data are missing; using a placeholder\cr
}
}
\section{Historical notes}{

\enumerate{
\item Until package version 0.1-14, released on 2025-02-10, longitude and
latitude were interpolated between surface values for the case where
\code{level} is given as 0. This behaviour was changed for issue 127, at
\url{https://github.com/dankelley/oceglider/issues/127}).
\item Until package version 0.1-16, released on 2025-02-15, data could be
erased for a while after power-on events, as controlled by a parameter
named \code{removeTimeSincePowerOn}. Now, a similar action may be accomplished
by calling \code{deleteStartupData()} on the return object from the present
function.
}
}

\section{Renaming of data in seaexplorer files}{


FIXME: are the original names the same in both "raw" and "sub" datasets?

Data in the \code{gli} files are stored in the \code{glider} item within
the \code{data} slot of the returned object, renamed as follows.
(Note that there is also an empty column in the seaexplorer data files,
caused by a semicolon at the ends of the lines. This is read by R, but
then discarded and not stored in the glider object.)

Data in the \code{gli} files are stored in the \code{glider} item within
the \code{data} slot of the returned object, renamed as follows. (If the
new name is listed as \code{.} that means that the old name is retained,
but bear in mind that a new name will likely be assigned at some later
point in the development of this package.)

\tabular{lll}{
\strong{OriginalName}       \tab \strong{Name}              \tab \strong{Notes}\cr
\code{Timestamp}            \tab \code{time}                \tab Converted to POSIXt time.\cr
\code{NavState}             \tab \code{navState}            \tab - \cr
\code{SecurityLevel}        \tab \code{alarm}               \tab - \cr
\code{Heading}              \tab \code{heading}             \tab - \cr
\code{Pitch}                \tab \code{pitch}               \tab - \cr
\code{Roll}                 \tab \code{roll}                \tab - \cr
\code{Depth}                \tab \code{pressureNav}         \tab - \cr
\code{Temperature}          \tab \code{temperatureInternal} \tab - \cr
\code{Pa}                   \tab \code{pressureInternal}    \tab - \cr
\code{Lat}                  \tab \code{latitude}            \tab Converted to decimal degrees.\cr
\code{Lon}                  \tab \code{longitude}           \tab Converted to decimal degrees.\cr
\code{DesiredH}             \tab \code{headingDesired}      \tab - \cr
\code{BallastCmd}           \tab \code{ballastCmd}          \tab - \cr
\code{BallastPos}           \tab \code{ballastPos}          \tab - \cr
\code{LinCmd}               \tab \code{linCmd}              \tab - \cr
\code{LinPos}               \tab \code{linPos}              \tab - \cr
\code{AngCmd}               \tab \code{angCmd}              \tab - \cr
\code{AngPos}               \tab \code{angPos}              \tab - \cr
\code{Voltage}              \tab \code{voltage}             \tab - \cr
\code{Altitude}             \tab \code{altitude}            \tab - \cr
}

Data in the \code{pld1} files are stored in the \code{payload1} item within
the \code{data} slot of the returned object, renamed as follows. Where possible,
the corresponding IOOS NetCDF file variable names are also listed [see
Integrated Ocean Observing System (U.S.). “NGDAC NetCDF File Format Version
2,” March 27, 2019.
https://github.com/ioos/ioosngdac/wiki/NGDAC-NetCDF-File-Format-Version-2].

\tabular{llll}{
\strong{OriginalName}       \tab \strong{Name}              \tab \strong{IOOS name} \tab \strong{Notes}\cr
\code{PLD_REALTIMECLOCK}    \tab \code{time}                \tab \code{time}        \tab Converted to POSIXt time.\cr
\code{NAV_RESOURCE}         \tab \code{navState}            \tab -                  \tab - \cr
\code{NAV_LONGITUDE}        \tab \code{longitude}           \tab \code{lon}         \tab Converted to decimal degrees.\cr
\code{NAV_LATITUDE}         \tab \code{latitude}            \tab \code{lat}         \tab Converted to decimal degrees.\cr
\code{NAV_DEPTH}            \tab \code{pressureNav}         \tab \code{depth}       \tab - \cr
\code{FLBBCD_CHL_COUNT}     \tab \code{chlorophyllCount}    \tab -                  \tab - \cr
\code{FLBBCD_CHL_SCALED}    \tab \code{chlorophyll}         \tab -                  \tab - \cr
\code{FLBBCD_BB_700_COUNT}  \tab \code{backscatterCount}    \tab -                  \tab - \cr
\code{FLBBCD_BB_700_SCALED} \tab \code{backscatter}         \tab -                  \tab - \cr
\code{FLBBCD_CDOM_COUNT}    \tab \code{cdomCount}           \tab -                  \tab - \cr
\code{FLBBCD_CDOM_SCALED}   \tab \code{cdom}                \tab -                  \tab - \cr
\code{GPCTD_CONDUCTIVITY}   \tab \code{conductivity}        \tab -                  \tab - \cr
\code{GPCTD_TEMPERATURE}    \tab \code{temperature}         \tab \code{temperature} \tab degC \cr
\code{GPCTD_PRESSURE}       \tab \code{pressure}            \tab \code{pressure}    \tab dbar \cr
\code{GPCTD_DOF}            \tab \code{oxygenFrequency}     \tab -                  \tab - \cr
-                           \tab \code{salinity}            \tab \code{salinity}    \tab Practical Salinity (computed)\cr
}

Data in the \code{pld2} files (or others for additional payloads)
are ignored in this version of the package. Please contact the
authors, if you need to handle such files.
}

\section{History of changes to this function}{

\itemize{
\item In version 0.1.16, the \code{removeTimeSincePowerOn} parameter was removed. The
new scheme is for the user to call \code{deleteStartupData()} on the return
value, to accomplish a similar thing.
}
}

\examples{
library(oceglider)

directory <- system.file("extdata/sea_explorer/delayed_raw",
    package =
        "oceglider"
)
g <- read.glider.seaexplorer.raw(directory, progressBar = FALSE)
plot(g, which = "p")

}
\seealso{
Other functions to read glider data: 
\code{\link{read.glider.netcdf}()},
\code{\link{read.glider.netcdf.ioos}()},
\code{\link{read.glider.slocum.csv}()},
\code{\link{read.glider.slocum.netcdf}()}
}
\author{
Clark Richards, Chantelle Layton and Dan Kelley
}
\concept{functions for seaexplorer gliders}
\concept{functions to read glider data}
