% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slocum_netcdf.R
\name{read.glider.slocum.netcdf}
\alias{read.glider.slocum.netcdf}
\title{Read a glider file in a Slocum netcdf format}
\usage{
read.glider.slocum.netcdf(
  file,
  readAll = FALSE,
  nameMap = list(salinity = "glider_record/sci_rbrctd_salinity_00", SA =
    "absolute_salinity", temperature = "glider_record/sci_rbrctd_temperature_00", CT =
    "conservative_temperature", pressure = "glider_record/sci_rbrctd_seapressure_00",
    time = "glider_record/sci_rbrctd_timestamp", conductivity =
    "glider_record/sci_rbrctd_conductivity_00", latitude = "glider_record/m_gps_lat",
    longitude = "glider_record/m_gps_lon", profileLat = "profile_lat", profileLon =
    "profile_lon", density = "density", oxygen = "oxygen_concentration", 
     u = "u", v
    = "v", vxi = "glider_record/m_initial_water_vx", vyi =
    "glider_record/m_initial_water_vy", vy = "glider_record/m_water_vy", vx =
    "glider_record/m_water_vx", id = "profile_id"),
  debug
)
}
\arguments{
\item{file}{Name of a netcdf file.}

\item{readAll}{logical value indicating whether to read all the data
columns in the file.  The default, FALSE, means to only
read columns that appear in \code{nameMap}.  Using FALSE can yield significant
decreases in processing time and memory usage.  For the 2 Gb test file
used during the coding of this function (which, admittedly, seems
to have a great deal of duplication or near-duplication of data),
setting \code{readAll} to FALSE drops the reading time from 12 s to 1.5s,
and the size of the resultant value from 2.0Gb to 0.28 Gb.}

\item{nameMap}{either a character value or a list.  In the first case,
the only permitted possibility is \code{nameMap="?"}, which instructs
\code{read.glider.slocum.netcdf()} to return a vector of variable names
as determined by scanning the file. Examining this list can be a useful
first step in the exploration of data file.  Once the variables are known, it makes
sense to use the list form of \code{nameMap}, perhaps starting with the default
value.}

\item{debug}{an integer controlling how much information is printed during
processing.  If this is 0, then only errors and warnings will be printed.  If it
is 1, then the function entry and exit are signalled, and a line
is printed for each variable read.}
}
\value{
A glider object, i.e. one inheriting from \linkS4class{glider}.
(This class inherits from \link[oce:oce-class]{oce::oce} in the
\CRANpkg{oce} package.)
}
\description{
This is a provisional function, written to handle a particular
file created in Spring of 2023 by MUN researchers, as part of an OTN
project.  Please consult the \dQuote{Slocum Gliders} vignette for more
information about this file.
}
\details{
Since different files might use different naming conventions for data,
it is helpful to establish a mapping between names in the file and the
names that are to be stored in the return object.  This is what the
\code{nameMap} argument is for.  The file may contain information that
will guide the user as to which of various possibilities to use,
although reference to other documents, on consultation with the
data provider may be required.  Again, see the \dQuote{Slocum Gliders}
vignette for more on this topic.
}
\seealso{
Other functions to read glider data: 
\code{\link{read.glider.netcdf.ioos}()},
\code{\link{read.glider.netcdf}()},
\code{\link{read.glider.seaexplorer.delayed}()},
\code{\link{read.glider.seaexplorer.realtime}()},
\code{\link{read.glider.slocum}()}

Other functions to read netcdf glider data: 
\code{\link{read.glider.netcdf.ioos}()},
\code{\link{read.glider.netcdf}()}
}
\author{
Dan Kelley, aided by Cameron Richardson on data names.
}
\concept{functions to read glider data}
\concept{functions to read netcdf glider data}
