% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{renameGlider}
\alias{renameGlider}
\title{Rename variables according to a specified dictionary}
\usage{
renameGlider(x, dictionary = "default", debug = 0)
}
\arguments{
\item{x}{either an \linkS4class{oce} object, the elements of which will
be renamed, or NULL. In the latter case, the dictionary is returned
as a data frame, which can be useful for users who want to use \code{\link[=rbind]{rbind()}}
to append dictionary elements of their own, thus customizing the
action of \code{rename()}.}

\item{dictionary}{either a string or a data frame.  If a string, then it is
either the name of a built-in vocabulary, either \code{ioos} or \code{sbe}
or the name of a CSV file that defines a dictionary in a four-column
format as described in \sQuote{Details}. If it is a data frame, then
it must hold four columns that follow the same pattern as in the CSV
style.}

\item{debug}{an integer, less than or equal to zero for no message, and
greater than zero for increasing levels of debugging.  Values greater than 4
are treated like 4.}
}
\description{
There are many conventions for naming oceanographic variables, and this
function provides a way to map names in data files to names to be used in an
object created from those files.
}
\details{
The dictionary format, whether read from a built-in CSV file, or from a
user-supplied CSV file, or as a data frame, contains four character-valued
columns, as follows.
\enumerate{
\item The original name of a variable in the \code{data} slot of \code{x}. This is used in
matching such names against targets. Matches may be in the form of
equality, or \link{regexp} match. In the latter case, a \verb{#} character may be
used as an abbreviation for a digit.  Note that \code{^} is inserted at the
start of the value, and \code{$} at the end, before searching for a match with
\code{\link[=grep]{grep()}}.
\item The desired oce-convention name to be used for a match. Many files will
yield duplicates, e.g. for multiple temperature sensors, so
\code{\link[=unduplicateNames]{unduplicateNames()}} is called after all names are processed, to avoid
problems.
\item The unit for the column, typically in a format handled by \code{\link[=expression]{expression()}}.
Note that this value is ignored if the object already holds stated units
for the quantity in question.
\item The scale for the column (again, only used if the object does not already
hold a scale).
}

The built-in dictionaries are stored in locations

\if{html}{\out{<div class="sourceCode">}}\preformatted{system.file("extdata", "dictionary_ioos.csv", package = "oceglider")
}\if{html}{\out{</div>}}

The data for these come from References 1.  The format
is simple, consisting of 4 columns, with no header. The column entries
are as follows.

\enumerate{

\item The first column holds a specialized regular expression for the variable
name as stored in the datafile.  This is conventional, except that \verb{#} is a
stand-in for the regular expression \verb{[0-9]} (that is, a single digit).
Formulating these expressions requires a bit of care, so it can make sense to
look at the \code{dictionary_sbe.csv} file to get some hints.

\item The second column holds the oce name.

\item The third column is the unit.

\item The fourth column is the scale.
}

In many cases, the third and fourth columns are empty, and even if values are
provided, they will be superceded by values within the data file.

As an example, the entry

\preformatted{
PSALST##,salinity,,PSS-78
}

indicates that a variable named \code{PSALT} followed by 2 digits is to be
renamed as \code{salinity}, that the unit (if not already defined within \code{x}) is
to be blank, and that the scale (again, if not already defined within \code{x}) is
to be \code{PSS-78}.
}
\references{
\enumerate{
\item IOOS naming convention
\url{https://cfconventions.org/Data/cf-standard-names/78/build/cf-standard-name-table.html}
}
}
\author{
Dan Kelley
}
