% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\name{setFlags,glider-method}
\alias{setFlags,glider-method}
\title{Set data-quality flags within a glider object}
\usage{
\S4method{setFlags}{glider}(object, name = NULL, i = NULL, value = NULL, debug = 0)
}
\arguments{
\item{object}{A glider object, i.e. an object inheriting from \code{glider-class}.}

\item{name}{Character string indicating the name of the variable to be flagged. If
this variable is not contained in the object's \code{data} slot, an error is reported.}

\item{i}{There are three choices for \code{i}. First, if
\code{i=="all"}, then any existing flags for the named item are discarded, and
replaced with the new \code{value}.  Second, if \code{i} is a vector of
integers, then flags are set to \code{value} at indices given by \code{i}.
Third, if it is a logical vector of the same length as the data, then just
those indices that match \code{TRUE} values in \code{i} are set to \code{value}.}

\item{value}{The value to be inserted in the flag.}

\item{debug}{Integer set to 0 for quiet action or to 1 for some debugging.}
}
\value{
An object with flags set as indicated.
}
\description{
This function changes specified entries in the data-quality
flags of \code{glider} objects. Those flags are stored within
a list named \code{flags$payload1} that resides in the \code{metadata}
slot.
}
\seealso{
See \code{\link[=handleFlags,glider-method]{handleFlags,glider-method()}} for an example of use.

Other functions relating to data-quality flags: 
\code{\link{handleFlags,glider-method}}
}
\author{
Dan Kelley
}
\concept{functions relating to data-quality flags}
