% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\docType{methods}
\name{summary,glider-method}
\alias{summary,glider-method}
\title{Summarize a glider Object}
\usage{
\S4method{summary}{glider}(object, ...)
}
\arguments{
\item{object}{A \code{glider} object, i.e. one inheriting from \code{glider-class}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarize a glider Object
}
