% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{swOxygenSolubility}
\alias{swOxygenSolubility}
\title{Compute Oxygen Solubility of Seawater}
\usage{
swOxygenSolubility(temperature, salinity)
}
\arguments{
\item{temperature}{numeric value holding temperature in degrees C.}

\item{salinity}{numeric value holding Practical Salinity.}
}
\description{
This function is a copy of Chantelle Layton's \code{swOxygen()} function,
from the \url{https://github.com/DFOglider/pilotingApp}
repository, renamed to avoid name-resolution conflicts.
}
\author{
Chantelle Layton
}
