% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{swSatO2}
\alias{swSatO2}
\title{Saturation of O2 in sea water}
\usage{
swSatO2(temperature, salinity)
}
\arguments{
\item{temperature}{temperature}

\item{salinity}{salinity

DEK note: the result is in mL/L.}
}
\description{
Computes the solubility (saturation) of Oxygen in sea water. Based on the
matlab function SW_SATO2 from the CSIRO seawater toolbox.
}
\author{
Chantelle Layton
}
