% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{toCamelCase}
\alias{toCamelCase}
\title{Convert a string from snake_case to camelCase}
\usage{
toCamelCase(s)
}
\arguments{
\item{s}{character value to be converted.}
}
\value{
CamelCase version of \code{s}.
}
\description{
Convert a string from snake_case to camelCase
}
\examples{
toCamelCase("profile_direction") # "profileDirection"

}
\author{
Dan Kelley
}
