% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceGlider.R
\name{urlExists}
\alias{urlExists}
\title{Check whether a URL exists, backtracing by separators, if not}
\usage{
urlExists(url, quiet = FALSE)
}
\arguments{
\item{url}{Character value specifying the URL. If no \code{/} is present
at the end of this string, then it is added before checks are done.}

\item{quiet}{Logical value indicating whether to print a suggestion
for an alternative website, in the case where \code{url} does not exist.}
}
\value{
A logical value indicating whether the website indicated
by \code{url} exists.
}
\description{
This uses \code{\link[RCurl:url.exists]{RCurl::url.exists()}} to see if the indicated URL exists.
If not, an attempt is made to find a lower-level URL that does exist.
This is done by progressively removing items separated by \code{"/"}
in \code{url}
}
\author{
Dan Kelley
}
