% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{degreeMinute}
\alias{degreeMinute}
\title{Convert lon and lat from a combined degree+minute formula}
\usage{
degreeMinute(x)
}
\arguments{
\item{x}{Numerical value in degree+minute notation (see \dQuote{Examples}).}
}
\value{
Numerical value in decimal degrees.
}
\description{
Data from Seaexplorers save longitude and latitude in a combined
format, in which e.g. 45deg 30.1min is saved as 4530.100, as
illustrated in the example.
}
\examples{
degreeMinute(4530.100) # 45+30.100/60

}
\author{
Dan Kelley
}
