% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{degreeMinute}
\alias{degreeMinute}
\title{Convert longitude and latitude from a combined degree+minute formula}
\usage{
degreeMinute(x)
}
\arguments{
\item{x}{Numerical value in degree+minute notation (see \dQuote{Examples}).}
}
\value{
\code{\link[=degreeMinute]{degreeMinute()}} returns a numerical value in decimal degrees.
}
\description{
SeaExplorer gliders save longitude and latitude with e.g.
4530.100 standing for 45deg 30.1min (which is 45.50167
in conventional decimal format).  This function converts SeaExplorer
coordinate values to conventional decimal values.
}
\examples{
degreeMinute(4530.100) # 45+30.100/60

}
\author{
Dan Kelley
}
