% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{getNextName}
\alias{getNextName}
\title{Find a Deconflicted Variable Name}
\usage{
getNextName(name, existingNames)
}
\arguments{
\item{name}{character value indicating the variable name.}

\item{existingNames}{vector of character values of existing
variable names.}
}
\value{
\link{getNextName} returns a deconflicted version of \code{name}.
}
\description{
This is used to name the first temperature found as \code{temperature},
the second as \code{temperature2}, etc.
}
\examples{
library(oceglider)
e <- NULL
e <- c(e, getNextName("S", e))
e <- c(e, getNextName("T", e))
e <- c(e, getNextName("S", e))
e <- c(e, getNextName("T", e))
e <- c(e, getNextName("p", e))
print(e)

}
\author{
Dan Kelley
}
