% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\docType{package}
\name{oceanglider-class}
\alias{oceanglider-class}
\alias{oceanglider-class-package}
\title{oceanglider: A Package for Processing Ocean Glider Data}
\description{
This package was written with two particular dataset types
in mind, from SeaExplorer and Slocum devices. There is a good
chance that the functions provided here (a) will fail on
other types and (b) function names and arguments will change
as more datasets are examined by the author.
}
