% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceglider.R
\name{plot,glider-method}
\alias{plot,glider-method}
\title{Plot an oceGlider Object}
\usage{
\S4method{plot}{glider}(x, which)
}
\arguments{
\item{x}{an \linkS4class{oceGlider} object.}

\item{which}{an integer FIXME: write more here}
}
\description{
FIXME: write more documentation here.
}
\author{
Dan Kelley
}
