% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{read.glider.netcdf}
\alias{read.glider.netcdf}
\title{Read a glider file in netcdf format}
\usage{
read.glider.netcdf(
  file,
  saveGlobalAttributes = TRUE,
  rename = TRUE,
  debug = getOption("gliderDebug", default = 0)
)
}
\arguments{
\item{file}{character value holding the name of a netcdf file
that holds glider data.}

\item{saveGlobalAttributes}{logical value indicating whether to
read the entirety of the global attributes stored within the
file into the \code{metadata} slot in a list named \code{globalAttributes}.}

\item{rename}{an indication of how (or if) to rename variables. This is
needed for most practical work in the oce package, which expects
standardized names, such as \code{"temperature"}, as opposed to the names stored
in glider files. There are three choices for \code{rename}. (a) It can be
logical, with TRUE (the default) meaning to use names as defined in
\code{system.file("extdata/dictionaries/seaexplorerDict.csv",package="oceglider")}
or FALSE, meaning not to rename variables. (b) It can be the name of a CSV
file that is in the same format as the file above-named file.  (c)
It can be a two-column data frame in which column 1 holds
variable names in the glider file and column 2 holds the corresponding
names to be used in the return value.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. If this is not provided, then
the value of \code{\link{getOption}("gliderDebug",0)} is used.
The printing is done by a call to \code{\link{gliderDebug}}.
Setting \code{debug=0} turns off this form of debugging, while
higher values may yield more information, depending on the
function. If one \code{glider} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
A glider object, i.e. one inheriting from \linkS4class{glider}.
(This class inherits from \link[oce:oce-class]{oce::oce} in the
\CRANpkg{oce} package.)
}
\description{
\strong{This is a provisional function, written to handle some
particular files available to the author.}
}
\details{
The data are copied directly from the file, except that \code{time}
is converted from an integer to a POSIX time. Variable names containing
underscores are renamed as e.g. \code{profile_direction}
to \code{profileDirection}, although the \code{\link{[[,glider-method}}
mechanism works with either name, e.g. if \code{g} is a glider object, then
\code{g[["profileDirection"]]} and
\code{g[["profile_direction"]]} give the same result.
}
\seealso{
Other functions to read glider data: 
\code{\link{read.glider.netcdf.ioos}()},
\code{\link{read.glider.seaexplorer.raw}()},
\code{\link{read.glider.slocum.csv}()},
\code{\link{read.glider.slocum.netcdf}()}

Other functions to read netcdf glider data: 
\code{\link{read.glider.netcdf.ioos}()},
\code{\link{read.glider.slocum.netcdf}()}
}
\author{
Dan Kelley
}
\concept{functions to read glider data}
\concept{functions to read netcdf glider data}
