% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\name{read.glider.netcdf}
\alias{read.glider.netcdf}
\title{Read a glider file in netcdf format}
\usage{
read.glider.netcdf(file)
}
\arguments{
\item{file}{Name of a netcdf file.}
}
\value{
A glider object, i.e. one inheriting from \code{\link{glider-class}}.
}
\description{
Read a glider file in netcdf format
}
\seealso{
Other functions to read glider data: \code{\link{read.glider.seaexplorer}},
  \code{\link{read.glider.slocum}}
}
\author{
Dan Kelley
}
\concept{functions to read glider data}
