% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\name{read.glider.netcdf}
\alias{read.glider.netcdf}
\title{Read a glider file in netcdf format}
\usage{
read.glider.netcdf(file, debug)
}
\arguments{
\item{file}{Name of a netcdf file.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. If this is not provided, then
the value of \code{\link{getOption}("gliderDebug",0)} is used.
The printing is done by a call to \code{\link{gliderDebug}}.
Setting \code{debug=0} turns off this form of debugging, while
higher values may yield more information, depending on the
function. If one \code{glider} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
A glider object, i.e. one inheriting from \code{\link{glider-class}}.
(This class inherits from \code{\link[oce]{oce-class}} in the
\code{oce} package.)
}
\description{
\strong{This is a provisional function, written to handle a particular level-1 file
provided to the author by DFO colleagues in mid January, 2019.} This only works
for files with global attribute \code{instrument} set to \code{Glider},
\code{instrument_manufacturer} set to \code{Alseamar}, and
\code{instrument_model} set to \code{SeaExplorer}, although this restriction
will likely be lifted as data from other instruments becomes available.
}
\details{
The data are copied directly from the file, except that \code{time}
is converted from an integer to a POSIX time. Variable names containing
underscores are renamed as e.g. \code{profile_direction}
to \code{profileDirection}, although the \code{\link{[[,glider-method}}
mechanism works with either name, e.g. if \code{g} is a glider object, then
\code{g[["profileDirection"]]} and
\code{g[["profile_direction"]]} give the same result.
}
\examples{
\dontrun{
library(oceanglider)
g <- read.glider.netcdf("GLI2018_SEA019_054DM_L1.nc")
## Remove spurious times (cannot be year 2009)
g <- subset(g, time > as.POSIXct("2018-01-01"))
## Remove bad data
g <- subset(g, is.finite(g[["salinity"]]))
## Focus on ascent phase (profileDirection==-1)
g <- subset(g, profileDirection==-1)
# CTD-style plot of whole dataset
ctd <- as.ctd(g[["salinity"]], g[["temperature"]], g[["pressure"]],
              longitude=g[["longitude"]], latitude=g[["latitude"]])
plot(ctd, type=rep("p", 4)) # 'type' gives dots
# CTD-style plot of a particular profile
plot(as.ctd(subset(g, profileIndex==200)))
}

}
\seealso{
Other functions to read glider data: \code{\link{read.glider.seaexplorer.delayed}},
  \code{\link{read.glider.seaexplorer.realtime}},
  \code{\link{read.glider.slocum}}
}
\author{
Dan Kelley
}
\concept{functions to read glider data}
