% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{read.glider}
\alias{read.glider}
\title{Read a glider data file}
\usage{
read.glider(file, debug, ...)
}
\arguments{
\item{file}{Character value giving the name of the file.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. If this is not provided, then
the value of \code{\link{getOption}("gliderDebug",0)} is used.
The printing is done by a call to \code{\link{gliderDebug}}.
Setting \code{debug=0} turns off this form of debugging, while
higher values may yield more information, depending on the
function. If one \code{glider} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}

\item{...}{Extra parameters passed to more specific \verb{read.*} functions.}
}
\value{
A \code{glider} object, i.e. one inheriting from \linkS4class{glider}.
}
\description{
This is a high-level function that passes control to \code{\link[=read.glider.netcdf]{read.glider.netcdf()}}
if the first argument is a string ending with \code{".nc"}, to
\code{\link[=read.glider.seaexplorer.realtime]{read.glider.seaexplorer.realtime()}} if it is a vector of strings, any
of which contains the text \code{"pld1.sub."} followed by one or more digits, or to
\code{\link[=read.glider.seaexplorer.delayed]{read.glider.seaexplorer.delayed()}} if it is a vector of strings, any
contains the text \code{"pld1.raw."} followed by one or more digits.
}
\author{
Dan Kelley
}
