% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaexplorer.R
\name{read.glider.seaexplorer.raw}
\alias{read.glider.seaexplorer.raw}
\title{Read raw SeaExplorer glider data}
\usage{
read.glider.seaexplorer.raw(dir, yo, level = 1, debug,
  progressBar = TRUE)
}
\arguments{
\item{dir}{The directory in which the raw SeaExplorer files are located.}

\item{yo}{A numeric value (or vector) specifying the yo numbers to
read. If not provided will read all yo numbers for which files
are present in \code{dir}.}

\item{level}{A numeric value specifying the processing level, 0 or
1. See Details.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. If this is not provided, then
the value of \code{\link{getOption}("gliderDebug",0)} is used.
The printing is done by a call to \code{\link{gliderDebug}}.
Setting \code{debug=0} turns off this form of debugging, while
higher values may yield more information, depending on the
function. If one \code{glider} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}

\item{progressBar}{A logical indicating whether to show progress
bars while reading the data. Can be useful when reading full
datasets.}
}
\description{
Reads raw CSV files produced by a SeaExplorer glider. The raw data
are the full resolution data stored on the glider and downloaded
after recovery.
}
\details{
While the data format is similar to the "real-time" SeaExplorer file
format, there are some differences and
\code{read.glider.explorer.realtime()} should be used for the latter.

This function can output either "Level 0" or "Level 1" type
data. Level 0 is simply the raw data as written in the CSV files
with no processing done, other than to remove longitude and
latitude for samples where the glider wasn't actually communicating
with satellites and then interpolate between surface values.

Level 1 processing performs a number of steps to give an
"analysis ready" dataset, including

\enumerate{

\item Interpolation of the surface longitude and latitude to give
an estimate of the subsurface positions. This is a crude estimate of
subsurface location and should be taken only as a first guess.

\item Removal of the first few sensor values from when the glider
is in \code{navState=118} (inflecting up) or \code{navState=110}
(inflecting down). The reason for this is that when the glider is
set to sample on alternating profiles, when the CTD is powered up
the first sample output to the payload computer is the \emph{last}
sample recorded before power down.

\item NAs for all the sensors are interpolated to a common
time. For example, if a Wetlabs FLBBCD sensor sampled, but there is
no corresponding GP-CTD sample from the same time, the CTD
parameters will be interpolated from the ones before and
after. This has the disadvantage of interpolating values that were
not measured, but has the advantage of assigning pressures to
values measured by sensors not integrated into the CTD
(e.g. Wetlabs FLBBCD, Rinko O2). Following the interpolation, any
rows with duplicated times are removed.

\item Calculate Practical salinity from conductivity, temperature
and pressure using \code{swSCTp()}.

}
}
\examples{
\dontrun{
library(oceanglider)
dir <- '/data/archive/glider/2019/sx/sea021m49/raw'
d <- read.glider.seaexplorer.raw(dir, yo=1:100)
plot(d, which=1)
}
}
\seealso{
Other functions for seaexplorer gliders: \code{\link{download.glider.seaexplorer}},
  \code{\link{read.glider.seaexplorer.realtime}}
}
\author{
Clark Richards and Dan Kelley
}
\concept{functions for seaexplorer gliders}
