% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaexplorer.R
\name{read.glider.seaexplorer}
\alias{read.glider.seaexplorer}
\title{Read a Seaexplorer Glider file}
\usage{
read.glider.seaexplorer(file, debug = 0)
}
\arguments{
\item{file}{File name.}

\item{debug}{Integer indicating the debugging level; 0 for quiet
action and higher values for more indications of the processing
steps.}
}
\description{
Read a Seaexplorer Glider file
}
\seealso{
Other functions for seaexplorer gliders: \code{\link{download.glider.seaexplorer}}

Other functions to read glider data: \code{\link{read.glider.slocum}}
}
\author{
Dan Kelley
}
\concept{functions for seaexplorer gliders}
\concept{functions to read glider data}
