% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slocum_netcdf.R
\name{read.glider.slocum.netcdf}
\alias{read.glider.slocum.netcdf}
\title{Read a glider file in a Slocum netcdf format}
\usage{
read.glider.slocum.netcdf(
  file,
  nameMap = list(salinity = "glider_record/sci_rbrctd_salinity_00", SA =
    "absolute_salinity", temperature = "glider_record/sci_rbrctd_temperature_00", CT =
    "conservative_temperature", pressure = "glider_record/sci_rbrctd_seapressure_00",
    time = "glider_record/sci_rbrctd_timestamp", conductivity =
    "glider_record/sci_rbrctd_conductivity_00", latitude = "glider_record/m_gps_lat",
    longitude = "glider_record/m_gps_lon", profileLat = "profile_lat", profileLon =
    "profile_lon", density = "density", O2 = "oxygen_concentration", 
     u = "u", v =
    "v", vxi = "glider_record/m_initial_water_vx", vyi =
    "glider_record/m_initial_water_vy", vy = "glider_record/m_water_vy", vx =
    "glider_record/m_water_vx", id = "profile_id"),
  debug
)
}
\arguments{
\item{file}{Name of a netcdf file.}

\item{nameMap}{List used to rename data columns. The default is tailored
to a particular file, and is subject to change without notice, as that
file is explored.}

\item{debug}{an integer controlling how much information is printed during
processing.  If this is 0, then only errors and warnings will be printed.  If it
is 1, then the function entry and exit are signalled, and a progress bar
is shown.  If it exceeds 1, then detailed notes are provided (but the
progress bar is not shown).}
}
\value{
A glider object, i.e. one inheriting from \linkS4class{glider}.
(This class inherits from \link[oce:oce-class]{oce::oce} in the
\CRANpkg{oce} package.)
}
\description{
\strong{This is an unsupported and provisional function}, written to handle a particular
file created in Spring of 2023 by MUN researchers, as part of an OTN
project.
}
\details{
The sample file does not use standard names for variables, but
the \code{nameMap} argument facilitates renaming for storage
in the returned object. (Renaming simplifies later analysis, e.g.
permitting direct use of algorithms in the \code{oce} package,
which assume that salinity is named \code{"salinity"}, etc.)
The original names of data items are retained in the metadata
of the returned object, so that the \code{[[} operator in the \code{oce}
package can retrieve the data using either the original name
or the more standard name.
}
\seealso{
Other functions to read glider data: 
\code{\link{read.glider.netcdf.ioos}()},
\code{\link{read.glider.netcdf}()},
\code{\link{read.glider.seaexplorer.delayed}()},
\code{\link{read.glider.seaexplorer.realtime}()},
\code{\link{read.glider.slocum}()}

Other functions to read netcdf glider data: 
\code{\link{read.glider.netcdf.ioos}()},
\code{\link{read.glider.netcdf}()}
}
\author{
Dan Kelley with help from Chantelle Layton and Cameron Richardson.
}
\concept{functions to read glider data}
\concept{functions to read netcdf glider data}
