% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slocum.R
\name{read.glider.slocum}
\alias{read.glider.slocum}
\title{Read a Slocum Glider file}
\usage{
read.glider.slocum(name)
}
\arguments{
\item{name}{Character value giving the name of the file}
}
\value{
An oce object holding the data.
}
\description{
Read a Slocum Glider file
}
\examples{
\dontrun{
g <- read.glider("~/data/glider/m80_2017-12-16_view_sci_water.csv")
}
}
\seealso{
Other functions for slocum gliders: \code{\link{download.glider.slocum}}
}
\author{
Dan Kelley
}
\concept{functions for slocum gliders}
\concept{functions to read glider data}
