% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\docType{methods}
\name{[[,glider-method}
\alias{[[,glider-method}
\title{Retrieve Part of a glider Object}
\usage{
\S4method{[[}{glider}(x, i, j, ...)
}
\arguments{
\item{x}{A glider object, i.e. one inheriting from \code{\link{glider-class}}.}

\item{i}{Character value that names the item to be retrieved.}

\item{j}{Optional character value specifying the data-stream to be used.}

\item{...}{Optional additional information (ignored).}
}
\description{
First, a check is done to see if the object's metadata contains an item
with name given by \code{i}. If this is true, then that value is returned.
Otherwise, the item is sought in the \code{data} slot. This is straightforward
for objects read by \code{\link{read.glider.slocum}}, which has no
repeated data items, but trickier for objects read by
\code{\link{read.glider.seaexplorer}}, since it stores data from the glider
and the payload separately, in items called \code{glider} and
\code{payload}, respectively. If \code{j} is not specified, then
\code{i} is sought first in the \code{payload} component, with
\code{glider} being checked thereafter. (For example, this means that the
payload thermometer is preferred to the glider thermometer.) This selection
process can be controlled by setting \code{j} to either \code{"glider"}
or \code{"payload"}.  For example, both \code{x[["temperature"]]} and
\code{x[["temperature","payload"]]} retrieves values from
the payload thermistor, while \code{x[["temperature","glider"]]} retrieve
values from the glider thermister.
}
\author{
Dan Kelley
}
