% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\docType{methods}
\name{subset,glider-method}
\alias{subset,glider-method}
\title{Subset an oceanglider Object}
\usage{
\S4method{subset}{glider}(x, subset, ...)
}
\arguments{
\item{x}{an oceanglider object, i.e. one inheriting from the \code{\link{glider-class}}.}

\item{subset}{a logical expression indicating how to take the subset.}

\item{...}{ignored}
}
\value{
An oceanglider object.
}
\description{
Note that \code{NA} values in the \code{subset} value will be dropped from
the return value, mimicking the behaviour of the base \code{link{subset}}
function.
}
\examples{
\dontrun{
# Example 1. remove wild salinities
library(oceanglider)
g <- read.glider(filename)
gg <- subset(g, 0 < salinity & salinity < 40)
par(mfrow=c(2, 1))
hist(g[["salinity"]], main="S original")
hist(gg[["salinity"]], main="S cleaned")
}

}
\author{
Dan Kelley
}
