% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanglider.R
\docType{methods}
\name{subset,glider-method}
\alias{subset,glider-method}
\title{Subset a glider Object}
\usage{
\S4method{subset}{glider}(x, subset, ...)
}
\arguments{
\item{x}{an oceanglider object, i.e. one inheriting from the \code{\link{glider-class}}.}

\item{subset}{a logical expression indicating how to take the subset. See \dQuote{Details}.}

\item{...}{ignored}
}
\value{
An oceanglider object.
}
\description{
Select a portion of an glider object, specified according to one
of two possible schemes, based on the form of the argument named
\code{subset}. Note that the schemes cannot be combined, so
nested calls must be used to accomplish combinations.
}
\details{
Scheme 1: if \code{subset} is a logical expression written
in terms of data that are stored in the yos, then this expression is applied
overall, to each 'yo' in the glider object (see Example 1).

Scheme 2: if \code{subset} is a logical expression containing
the word \code{"levels"}, then the expression is used as a filter
to select yos (see Example 2). Typically, this might be used to avoid
short yos. If \code{subset} is the word \code{"ascending"}, then
only ascending segments of yos are retained, while if it is
\code{"descending"} then only descending segments are retained. These
schemes only work for seaexplorer data, and they are based on filtering
by \code{navState} (with 117 meaning an ascending portion and 100
meaning a descending portion).
}
\examples{
\dontrun{

# Example 1. remove wild salinities
library(oceanglider)
g <- read.glider(filename)
gg <- subset(g, 0 < salinity & salinity < 40)
par(mfrow=c(2, 1))
hist(g[["salinity"]], main="S original")
hist(gg[["salinity"]], main="S cleaned")

# Example 2. remove short yos
gg <- subset(g, levels > 4)

# Example 3. retain only ascending portions of yos
gascending <- subset(g, "ascending")

# Example 4. retain only descending portions of yos
gdescending <- subset(g, "descending")
}

}
\author{
Dan Kelley
}
