% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{swOxygenFrequencyToSaturation}
\alias{swOxygenFrequencyToSaturation}
\title{Convert SBE Oxygen Frequency Value to Oxygen Saturation}
\usage{
swOxygenFrequencyToSaturation(
  temperature,
  salinity,
  pressure,
  frequency,
  cal,
  unit = "ml/l"
)
}
\arguments{
\item{temperature}{numeric value holding temperature in degrees C.}

\item{salinity}{numeric value holding Practical Salinity.}

\item{pressure}{numeric value holding pressure in dbar.}

\item{frequency}{the frequency reading from the oxygen sensor
system, in Hz.}

\item{cal}{list or data frame holding calibration parameters named
`Soc`, `Foffset`, `A`, `B`, `C`, `D` and `Enom`.  The last of these
is called "E nominal" in the calibration sheets.}

\item{unit}{character value indicating the desired unit. The
default is `ml/l` but `umol/kg` is also accepted, the latter being
computed by multiplying the former by 44.6591. (FIXME: document
the source of this number.)}
}
\description{
The calibration formula comes from SBE calibration sheets. This
code is based on `sbeO2Hz2Sat()` in
\url{https://github.com/DFOglider/pilotingApp}, with minor
modifiations: (1) it has a new name to avoid conflicts, (2) the
calibration coefficients are collected into a single parameter,
`cal`, (3) `oxygenFrequency` is renamed `frequency`, and (4) `unit`
is added, to permit specifying the unit of the return value.
}
\author{
Chantelle Layton and Dan Kelley
}
