% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{toCamelCase}
\alias{toCamelCase}
\title{Convert a string from snake_case to camelCase}
\usage{
toCamelCase(s)
}
\arguments{
\item{s}{character value to be converted.}
}
\value{
\code{\link[=toCamelCase]{toCamelCase()}} returns a camelCase version of \code{s}.
}
\description{
Convert a snake-case string (i.e., one constructed
with words separated by underlines) to a camel-case
string (i.e. one in which the words are strung together,
with upper-case for the first letter of all but the
first word).  See \sQuote{Examples}.
}
\examples{
toCamelCase("profile_direction") # "profileDirection"

}
\author{
Dan Kelley
}
